Refactored code by separating game logic functions into the Game struct and moving other structs to separate files for better organization and modularity.
